requirements.txt
requests~=2.28
jsonpath-ng~=1.5
python-dotenv~=1.0
tabulate~=0.9.0
